pyproject.toml
[tool.poetry]
name = "advertisement-sevice"
version = "0.1.0"
description = ""
authors = ["SemyonF <Falaleevmay@gmail.com>"]
readme = "README.md"
packages = [
    { include = "src", from = "." }
]

[tool.poetry.dependencies]
python = ">=3.12"
sqlalchemy = "^2.0.40"
fastapi = "^0.115.12"
asyncpg = "^0.30.0"
alembic = "^1.15.2"
uvicorn = "^0.34.0"
passlib = "^1.7.4"
pyjwt = "^2.10.1"
bcrypt = "^4.3.0"
python-multipart = "^0.0.20"
aiogram = "^3.19.0"
pydantic = {version = ">=2.4.1,<2.11", extras = ["email"]}
httpx = "^0.28.1"
black = "^25.1.0"
async-factory-boy = "^1.0.1"
pytest-asyncio = "^0.26.0"
pytest = "^8.3.5"
python-dotenv = "^1.1.0"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"